Add new files

+++ delete.php
@@ -1,11 +1,20 @@
 <?php
 
-$connect = mysqli_connect('localhost', 'root', '', 'test');
-if ($connect) {
-} else {
-    die('Błąd połączenia z bazą danych');
-}
+include('sql.php');
+
 $surname = $_POST['surname'];
-$sql = "DELETE FROM reserv WHERE surname = $surname ";
-mysqli_query($connect, $sql);
-header('Location: admin.php');
+
+$s = "SELECT * FROM reserv WHERE surname='$username';";
+
+$result = mysqli_query($connect, $s);
+
+$num = mysqli_num_rows($result);
+
+
+if ($num == 1) {
+    header('Location: admin.php');
+} else {
+    $sql = "DELETE FROM reserv WHERE surname = $surname ";
+    mysqli_query($connect, $sql);
+    header('Location: admin.php');
+}
